requirements.txt
pyyaml
shapely
ipython
descartes
networkx
matplotlib
subprocess32
